refactor: app.js 삭제하고 app 함수 index.js로 이동
