--- src/js/app.js
@@ -1,15 +1,0 @@
-import Header from './components/header/index.js';
-import Nav from './components/nav/index.js';
-import MainContent from './components/mainContent/index.js';
-
-const app = async ($targetEle) => {
-  const header = new Header($targetEle);
-  const nav = new Nav($targetEle);
-  const mainContent = new MainContent($targetEle);
-
-  header.mount();
-  await nav.mount();
-  mainContent.mount();
-};
-
-export default app;
